Update UserController to include logout functionality
This commit adds a new method `logout` to the `UserController` class. The `logout` method logs out the authenticated user by calling the `auth()->logout()` function. After logging out, the user is redirected to the homepage with a success message. This change allows users to easily log out of the application.